cups: Add driver info for Phomemo M421
Phomemo M421 is a larger variant of a Phomemo M220 label printer.  It is capable
of printing on 4x6in media (common size of shipping labels in the US).
According to the spec, this printer supports labels with a width of 40-106mm.

This commit adds a separate driver info file for M421 which is similar to M220,
but has a different set of supported media. The list of supported media matches
what is currently available on Phomemo website for this printer.